Client held item change converted
