+++ src/main/kotlin/packets/play/in/ClientCreativeInventoryActionPacket.kt
@@ -36,20 +36,4 @@ class ClientCreativeInventoryActionPacket(data: ByteArray) : Packet(data) {
             }
         }
     }
-
-    private fun sendHeldItemUpdate(client: ClientSession) {
-        val stack = client.player.getHeldItem()
-        val slotData = stack.toSlotData()
-
-        for(otherPlayer in Bullet.players) {
-            if(otherPlayer != client.player) {
-                otherPlayer.sendPacket(
-                    ServerEntityEquipmentPacket(
-                        client.player.entityID,
-                        listOf(0 to slotData)
-                    )
-                )
-            }
-        }
-    }
 }
