+++ src/main/kotlin/packets/play/in/ClientHeldItemChangePacket.kt
@@ -1,5 +1,8 @@
 package com.aznos.packets.play.`in`
 
+import com.aznos.ClientSession
+import com.aznos.events.EventManager
+import com.aznos.events.PlayerHeldItemChangeEvent
 import com.aznos.packets.Packet
 
 /**
@@ -7,4 +10,13 @@ import com.aznos.packets.Packet
  */
 class ClientHeldItemChangePacket(data: ByteArray) : Packet(data) {
     val slot: Short = getIStream().readShort()
+
+    override fun apply(client: ClientSession) {
+        val event = PlayerHeldItemChangeEvent(client.player, slot.toInt())
+        EventManager.fire(event)
+        if(event.isCancelled) return
+
+        client.player.selectedSlot = slot.toInt()
+        client.player.sendHeldItemUpdate(client)
+    }
 }
